requirements.txt
python-telegram-bot==21.6
pandas==2.1.4
google-api-python-client==2.147.0
google-auth==2.35.0
apscheduler==3.10.4
pytz==2024.1
